Add another example of service: InventoryService, and move ItemFactory to core.items package
