+++ src/main/kotlin/core/items/InventoryRepository.kt
@@ -1,0 +1,5 @@
+package dev.deadzone.core.items
+
+interface InventoryRepository {
+    fun placeholder()
+}
