+++ src/main/kotlin/core/items/InventoryRepositoryMongo.kt
@@ -1,0 +1,7 @@
+package dev.deadzone.core.items
+
+class InventoryRepositoryMongo: InventoryRepository {
+    override fun placeholder() {
+
+    }
+}
