+++ src/main/kotlin/core/items/InventoryService.kt
@@ -1,0 +1,11 @@
+package dev.deadzone.core.items
+
+import dev.deadzone.core.PlayerService
+
+class InventoryService(
+    private val inventoryRepository: InventoryRepository
+) : PlayerService {
+    override suspend fun init(playerId: String) {
+
+    }
+}
