+++ src/main/kotlin/socket/handler/saveresponse/crate/CrateUnlockResponse.kt
@@ -1,6 +1,6 @@
 package dev.deadzone.socket.handler.saveresponse.crate
 
-import dev.deadzone.core.factory.ItemFactory
+import dev.deadzone.core.items.ItemFactory
 import dev.deadzone.core.model.game.data.Item
 import kotlinx.serialization.Serializable
